+++ ribs/emitters/opt/_cma_es.py
@@ -133,11 +133,6 @@ class CMAEvolutionStrategy(EvolutionStrategyBase):
         self.cov = None
 
     def reset(self, x0):
-        """Resets the optimizer to start at x0.
-
-        Args:
-            x0 (np.ndarray): Initial mean.
-        """
         self.current_eval = 0
         self.sigma = self.sigma0
         self.mean = np.array(x0, self.dtype)
@@ -150,18 +145,7 @@ class CMAEvolutionStrategy(EvolutionStrategyBase):
         self.cov = DecompMatrix(self.solution_dim, self.dtype)
 
     def check_stop(self, ranking_values):
-        """Checks if the optimization should stop and be reset.
-
-        Tolerances come from CMA-ES.
-
-        Args:
-            ranking_values (np.ndarray): Array of objective values of the
-                solutions, sorted in the same order that the solutions were
-                sorted when passed to ``tell()``.
-
-        Returns:
-            True if any of the stopping conditions are satisfied.
-        """
+        # Tolerances from pycma CMA-ES.
         if self.cov.condition_number > 1e14:
             return True
 
@@ -198,12 +182,6 @@ class CMAEvolutionStrategy(EvolutionStrategyBase):
     # multithreading because our data is too small.
     @threadpool_limits.wrap(limits=1, user_api="blas")
     def ask(self, batch_size=None):
-        """Samples new solutions from the Gaussian distribution.
-
-        Args:
-            batch_size (int): batch size of the sample. Defaults to
-                ``self.batch_size``.
-        """
         if batch_size is None:
             batch_size = self.batch_size
 
@@ -266,15 +244,7 @@ class CMAEvolutionStrategy(EvolutionStrategyBase):
     # Limit OpenBLAS to single thread. This is typically faster than
     # multithreading because our data is too small.
     @threadpool_limits.wrap(limits=1, user_api="blas")
-    def tell(self, ranking_indices, num_parents):
-        """Passes the solutions back to the optimizer.
-
-        Args:
-            ranking_indices (array-like of int): Indices that indicate the
-                ranking of the original solutions returned in ``ask()``.
-            num_parents (int): Number of top solutions to select from the
-                ranked solutions.
-        """
+    def tell(self, ranking_indices, ranking_values, num_parents):
         self.current_eval += len(self._solutions[ranking_indices])
 
         if num_parents == 0: