+++ HISTORY.md
@@ -7,6 +7,8 @@
 #### API
 
 - Add PyCMAEvolutionStrategy for using pycma in ES emitters ({pr}`434`)
+- **Backwards-incompatible:** Add ranking values to evolution strategy tell
+  method ({pr}`438`)
 - **Backwards-incompatible:** Move evolution strategy bounds to init ({pr}`436`)
 - **Backwards-incompatible:** Use seed instead of rng in ranker ({pr}`432`)
 - **Backwards-incompatible:** Replace status and value with add_info ({pr}`430`)